Bump packages versions and fix safehts test

transfer to one file
