+++ package.json
@@ -17,11 +17,11 @@
   },
   "homepage": "https://github.com/hashgraph/hedera-smart-contracts#readme",
   "devDependencies": {
-    "@hashgraph/hedera-local": "^1.0.2",
+    "@hashgraph/hedera-local": "^2.3.2",
     "@nomicfoundation/hardhat-chai-matchers": "^1.0.5",
     "@openzeppelin/contracts": "^4.8.0",
     "@openzeppelin/contracts-upgradeable": "^4.8.0",
-    "@openzeppelin/hardhat-upgrades": "^1.21.0",
+    "@openzeppelin/hardhat-upgrades": "^1.22.0",
     "hardhat": "^2.12.4"
   },
   "dependencies": {
